Move profile to it's own file
& standardized questions
