+++ tdx/utils.py
@@ -2,7 +2,7 @@ from typing import Union
 
 import discord
 from redbot.core.i18n import Translator
-from redbot.core.utils import chat_formatting as cf, predicates
+from redbot.core.utils import chat_formatting as cf
 
 from . import client
 
@@ -15,22 +15,17 @@ def check_xp(x: client.Update) -> int:
     return x.total_xp
 
 
-def contact_us_on_twitter() -> str:
-    return cf.info(
-        _("If that doesn't look right, please contact us on Twitter. {twitter_handle}")
-    ).format(twitter_handle="@TrainerDexApp")
+TWITTER_MESSAGE = cf.info(
+    _("If that doesn't look right, please contact us on Twitter. {twitter_handle}")
+).format(twitter_handle="@TrainerDexApp")
 
 
 def append_twitter(text: str) -> str:
-    return "{original_message}\n\n{twitter}".format(
-        original_message=text, twitter=contact_us_on_twitter()
-    )
+    return f"{text}\n\n{TWITTER_MESSAGE}"
 
 
 def append_icon(icon: Union[discord.Emoji, str], text: str) -> str:
-    if icon:
-        return f"{icon} {text}"
-    return text
+    return f"{icon} {text}" if icon else text
 
 
 def quote(text: str) -> str:
@@ -62,36 +57,6 @@ def success(text: str) -> str:
     """
     emoji = "\N{WHITE HEAVY CHECK MARK}"
     return append_icon(emoji, text)
-
-
-class QuestionMessage:
-    def __init__(self, ctx, question: str, check=None):
-        self._ctx = ctx
-        self.question = question
-        self.message = None
-        if check:
-            self.predicate = check
-        else:
-            self.predicate = predicates.MessagePredicate.same_context(self._ctx)
-        self.response = None
-
-    async def ask(self, bot):
-        self.message = await self._ctx.send(
-            cf.question(f"{self._ctx.author.mention}: {self.question}")
-        )
-        self.response = await bot.wait_for("message", check=self.predicate)
-
-    @property
-    def answer(self) -> str:
-        if self.response:
-            return self.response.content
-        return None
-
-    @property
-    def exit(self) -> bool:
-        if self.answer:
-            return self.answer.lower() == f"{self._ctx.prefix}cancel"
-        return False
 
 
 def introduction_notes(ctx, member, trainer, additional_message: str) -> str:
